python/requirements_mac.txt
numpy>=1.12, <=1.16.4 ; python_version<"3.5"
shapely
wheel>=0.34.0, <0.35.0
setuptools>=44.1.0
google>=2.0.3
protobuf>=3.12.2
func-timeout>=4.3.5
pyyaml>=5.1
flask>=1.1.2
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
ujson>=2.0.3
grpcio-tools
grpcio>=1.33.2
sentencepiece==0.1.96; platform_machine !=  "aarch64"
sentencepiece; platform_machine ==  "aarch64"
opencv-python==3.4.17.61; platform_machine !=  "aarch64"
opencv-python; platform_machine == "aarch64"
pillow==8.4.0 ; python_version == "3.6"
pillow==9.0.0 ; python_version > "3.6"

